+++ tests/TestHelper/Container/ClassD.php
@@ -9,7 +9,7 @@
  */
 declare(strict_types=1);
 
-namespace Linna\TestHelper\DI;
+namespace Linna\TestHelper\Container;
 
 class ClassD
 {