+++ tests/TestHelper/Container/ClassF.php
@@ -9,9 +9,9 @@
  */
 declare(strict_types=1);
 
-namespace Linna\TestHelper\DI;
+namespace Linna\TestHelper\Container;
 
-class ClassI
+class ClassF
 {
     public function __construct()
     {