+++ tests/TestHelper/Container/ClassI.php
@@ -9,9 +9,9 @@
  */
 declare(strict_types=1);
 
-namespace Linna\TestHelper\DI;
+namespace Linna\TestHelper\Container;
 
-class ClassH
+class ClassI
 {
     public function __construct()
     {